LPD-4254 - reorganize previous test and create a new one of the new behavaior
