+++ modules/test/playwright/playwright.config.ts
@@ -99,7 +99,8 @@ import {config as scimConfiguraitonWebConfig} from './tests/scim-configuration-w
 import {config as searchExperiencesWebConfig} from './tests/search-experiences-web/config';
 import {config as segmentExperimentWebConfig} from './tests/segment-experiment-web/config';
 import {config as segmentsWebConfig} from './tests/segments-web/config';
-import {config as setAdminPassword} from './tests/set-admin-password/config';
+import {config as setAdminPasswordBlank} from './tests/set-admin-password/set-admin-password-blank/config';
+import {config as setAdminPasswordTest} from './tests/set-admin-password/set-admin-password-test/config';
 import {
 	pageManagementSiteSetup,
 	pageManagementSiteTeardown,
@@ -224,7 +225,8 @@ export default defineConfig({
 		searchExperiencesWebConfig,
 		segmentExperimentWebConfig,
 		segmentsWebConfig,
-		setAdminPassword,
+		setAdminPasswordBlank,
+		setAdminPasswordTest,
 		siteAdminWebConfig,
 		siteNavigationAdminWebConfig,
 		siteNavigationBreadcrumbWebConfig,
