+++ modules/test/playwright/tests/set-admin-password/set-admin-password-blank/config.ts
@@ -1,0 +1,9 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+export const config = {
+	name: 'set-admin-password-blank',
+	testDir: 'tests/set-admin-password/set-admin-password-blank',
+};
